requirements.txt
psutil
pystray
Pillow
plyer
